Rewrite of farm plot and crop code
This one's a doozy. Basically I rewrote the farm plot code so that the
plot and the crops are decoupled, and so that the crops actually use
growth stages. All of this is done using the AutomataGrid datatype that
I created from the TM_Manager. AutomataGrid is agnostic of what it
stores, so the crops and plot interpret and set it up. Theoretically
more crops may be added by simply adding to the enum it pulls from in
setup (currently in Seedbag, to be moved soon), though it will not
handle the crops in the second column without modification (though the
modification should be relatively slight?)